// dashboard/ui/src/api/generated/models/createExperimentForm.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */
import type { CreateExperimentFormType } from './createExperimentFormType.ts';
import type { Engine } from './engine.ts';

export interface CreateExperimentForm {
  /** @minLength 1 */
  'experiment-name': string;
  type: CreateExperimentFormType;
  pdb?: string;
  'repo-url'?: string;
  'notebooks-repo'?: string;
  'access-token'?: string;
  'notebook-module'?: string;
  engine?: Engine;
  'simulation-files'?: Blob[];
}